Changed the way invalidation reference is generated (#21)
## Motivation
Every path that needs to be invalidated is a different entry in the queue and when two messages arrive to the lambda at the same very microsecond, the uniqueness that is provided by the timestamp is not enough. Instead of relying on that, the unique message id can be used.